refactor(jni): functions reordering
