+++ zenoh-jni/src/queryable.rs
@@ -15,22 +15,24 @@
 use std::sync::Arc;
 
 use jni::{
-    objects::{JClass, JObject, JString},
-    sys::jboolean,
+    objects::{GlobalRef, JByteArray, JClass, JObject, JString, JValue},
+    sys::{jboolean, jint, jlong},
     JNIEnv,
 };
-use zenoh::queryable::Queryable;
 use zenoh::{
     key_expr::KeyExpr,
     prelude::Wait,
     session::{Session, SessionDeclarations},
 };
+use zenoh::{query::Query, queryable::Queryable};
 
 use crate::{
     errors::{Error, Result},
     key_expr::process_kotlin_key_expr,
-    query::on_query,
-    utils::{get_callback_global_ref, get_java_vm, load_on_close},
+    utils::{
+        bytes_to_java_array, get_callback_global_ref, get_java_vm, load_on_close,
+        slice_to_java_string,
+    },
 };
 
 /// Frees the memory associated with a Zenoh queryable raw pointer via JNI.
@@ -119,4 +121,112 @@ pub(crate) unsafe fn declare_queryable(
     queryable
         .wait()
         .map_err(|err| Error::Session(format!("Error declaring queryable: {}", err)))
+}
+
+/// Handles a Zenoh query callback via JNI.
+///
+/// This function is responsible for invoking the query callback function provided by the user from Java/Kotlin.
+///
+/// Parameters:
+/// - `env`: The JNI environment.
+/// - `query`: The Zenoh [Query] to be passed to the callback function.
+/// - `callback_global_ref`: A global object reference of the callback function.
+///
+/// Returns:
+/// - A [Result] indicating success or failure.
+///
+/// Note:
+/// - The callback reference `callback_global_ref` should point to the Java/Kotlin implementation
+///   of the `onQuery` function (which receives a `Query` as a parameter) from the `Callback` interface.
+///
+pub(crate) fn on_query(
+    mut env: JNIEnv,
+    query: Query,
+    callback_global_ref: &GlobalRef,
+) -> Result<()> {
+    let selector = query.selector();
+    let value = query.value();
+
+    let selector_params_jstr =
+        env.new_string(selector.parameters().to_string())
+            .map_err(|err| {
+                Error::Jni(format!(
+                    "Could not create a JString through JNI for the Query key expression. {}",
+                    err
+                ))
+            })?;
+
+    let (with_value, payload, encoding_id, encoding_schema) = if let Some(value) = value {
+        let byte_array = bytes_to_java_array(&env, value.payload())?;
+        let encoding_id = value.encoding().id() as jint;
+        let encoding_schema = match value.encoding().schema() {
+            Some(schema) => slice_to_java_string(&env, schema)?,
+            None => JString::default(),
+        };
+        (true, byte_array, encoding_id, encoding_schema)
+    } else {
+        (false, JByteArray::default(), 0, JString::default())
+    };
+
+    let attachment_bytes = query
+        .attachment()
+        .map_or_else(
+            || Ok(JByteArray::default()),
+            |attachment| bytes_to_java_array(&env, attachment),
+        )
+        .map_err(|err| Error::Jni(format!("Error processing attachment of reply: '{}'.", err)))?;
+
+    let key_expr_str = env
+        .new_string(&query.key_expr().to_string())
+        .map_err(|err| {
+            Error::Jni(format!(
+                "Could not create a JString through JNI for the Query key expression. {}",
+                err
+            ))
+        })?;
+
+    let query_ptr = Arc::into_raw(Arc::new(query));
+
+    let result = env
+        .call_method(
+            callback_global_ref,
+            "run",
+            "(Ljava/lang/String;Ljava/lang/String;Z[BILjava/lang/String;[BJ)V",
+            &[
+                JValue::from(&key_expr_str),
+                JValue::from(&selector_params_jstr),
+                JValue::from(with_value),
+                JValue::from(&payload),
+                JValue::from(encoding_id),
+                JValue::from(&encoding_schema),
+                JValue::from(&attachment_bytes),
+                JValue::from(query_ptr as jlong),
+            ],
+        )
+        .map(|_| ())
+        .map_err(|err| {
+            // The callback could not be invoked, therefore the created kotlin query object won't be
+            // used. Since `query_ptr` as well as `key_expr_ptr` was created within this function
+            // and remains unaltered, it is safe to reclaim ownership of the memory by converting
+            // the raw pointers back into an `Arc` and freeing the memory.
+            unsafe {
+                Arc::from_raw(query_ptr);
+            };
+            _ = env.exception_describe();
+            Error::Jni(format!("{}", err))
+        });
+
+    _ = env
+        .delete_local_ref(key_expr_str)
+        .map_err(|err| tracing::error!("Error deleting local ref: {}", err));
+    _ = env
+        .delete_local_ref(selector_params_jstr)
+        .map_err(|err| tracing::error!("Error deleting local ref: {}", err));
+    _ = env
+        .delete_local_ref(payload)
+        .map_err(|err| tracing::error!("Error deleting local ref: {}", err));
+    _ = env
+        .delete_local_ref(attachment_bytes)
+        .map_err(|err| tracing::error!("Error deleting local ref: {}", err));
+    result
 }
